+++ legend-engine-language-pure-compiler/src/main/java/org/finos/legend/engine/language/pure/compiler/toPureGraph/MilestoningDatePropagationHelper.java
@@ -14,55 +14,73 @@
 
 package org.finos.legend.engine.language.pure.compiler.toPureGraph;
 
+import org.eclipse.collections.api.RichIterable;
 import org.eclipse.collections.api.list.ImmutableList;
 import org.eclipse.collections.api.list.ListIterable;
+import org.eclipse.collections.api.list.MutableList;
+import org.eclipse.collections.api.tuple.Pair;
 import org.eclipse.collections.impl.factory.Lists;
+import org.eclipse.collections.impl.list.mutable.FastList;
+import org.eclipse.collections.impl.tuple.Tuples;
+import org.eclipse.collections.impl.utility.ArrayIterate;
+import org.eclipse.collections.impl.utility.LazyIterate;
+import org.eclipse.collections.impl.utility.ListIterate;
+import org.finos.legend.engine.protocol.pure.v1.model.SourceInformation;
+import org.finos.legend.engine.protocol.pure.v1.model.context.EngineErrorType;
 import org.finos.legend.engine.protocol.pure.v1.model.valueSpecification.application.AppliedFunction;
 import org.finos.legend.engine.shared.core.operational.Assert;
+import org.finos.legend.engine.shared.core.operational.errorManagement.EngineException;
 import org.finos.legend.pure.m3.compiler.postprocessing.processor.milestoning.MilestoningStereotype;
+import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.extension.Stereotype;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.function.property.AbstractProperty;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.function.property.QualifiedProperty;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.type.Class;
+import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.type.Type;
+import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.valuespecification.FunctionExpression;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.valuespecification.SimpleFunctionExpression;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.valuespecification.ValueSpecification;
+import org.finos.legend.pure.m3.navigation.profile.Profile;
+
+import java.util.Objects;
 
 public class MilestoningDatePropagationHelper {
 
     private static final String ALL_VERSIONS_IN_RANGE_PROPERTY_NAME_SUFFIX = "AllVersionsInRange";
     private static final String ALL_VERSIONS_PROPERTY_NAME_SUFFIX = "AllVersions";
+    private static final ImmutableList<String> MILESTONING_DATE_SOURCE_TYPES = Lists.immutable.of("getAll", "exists", "project", "filter", "subType", "map");
 
-    public static boolean checkGetAllFunctionWithMilestoningContext(SimpleFunctionExpression func)
+    public static boolean isGetAllFunctionWithMilestoningContext(SimpleFunctionExpression func)
     {
        return "getAll".equals(func._functionName()) && ("getAll_Class_1__Date_1__T_MANY_".equals(func._func().getName()) || "getAll_Class_1__Date_1__Date_1__T_MANY_".equals(func._func().getName()));
     }
 
     public static void setMilestoningPropagationContext(SimpleFunctionExpression func, ProcessingContext processingContext)
     {
-        Assert.assertTrue(checkGetAllFunctionWithMilestoningContext(func), () -> "MilestoneDatePropagationContext can only be set from getAll() function");
+        Assert.assertTrue(isGetAllFunctionWithMilestoningContext(func), () -> "MilestoneDatePropagationContext can only be set from getAll() function");
         ValueSpecification parameterValue = func._parametersValues().getFirst();
-        MilestoningStereotype milestoningStereotype = Milestoning.temporalStereotypes(parameterValue._genericType()._typeArguments().toList().getFirst()._rawType()._stereotypes()).get(0);
+        MilestoningStereotype milestoningStereotype = Milestoning.temporalStereotypes(parameterValue._genericType()._typeArguments().toList().getFirst()._rawType()._stereotypes());
         ListIterable<? extends ValueSpecification> temporalParameterValues = func._parametersValues().toList();
         setMilestoningDates(temporalParameterValues, milestoningStereotype, processingContext);
     }
 
     public static void setMilestoningDates(ListIterable<? extends ValueSpecification> temporalParameterValues, MilestoningStereotype milestoningStereotype, ProcessingContext processingContext)
     {
-        if (Milestoning.isBusinessTemporal(milestoningStereotype))
+        if (isBusinessTemporal(milestoningStereotype))
         {
             processingContext.milestoningDatePropagationContext.setBusinessDate(temporalParameterValues.get(1));
         }
-        else if (Milestoning.isProcessingTemporal(milestoningStereotype))
+        else if (isProcessingTemporal(milestoningStereotype))
         {
             processingContext.milestoningDatePropagationContext.setProcessingDate(temporalParameterValues.get(1));
         }
-        else if (Milestoning.isBiTemporal(milestoningStereotype))
+        else if (isBiTemporal(milestoningStereotype))
         {
             processingContext.milestoningDatePropagationContext.setProcessingDate(temporalParameterValues.get(1));
             processingContext.milestoningDatePropagationContext.setBusinessDate(temporalParameterValues.get(2));
         }
     }
 
-    public static boolean checkForFilter(SimpleFunctionExpression func)
+    public static boolean isFilter(SimpleFunctionExpression func)
     {
         return "filter".equals(func._functionName());
     }
@@ -72,7 +90,7 @@ public class MilestoningDatePropagationHelper {
         if(!property._func()._functionName().endsWith(ALL_VERSIONS_IN_RANGE_PROPERTY_NAME_SUFFIX) && !property._func()._functionName().endsWith(ALL_VERSIONS_PROPERTY_NAME_SUFFIX))
         {
             Class target = (Class) property._genericType()._rawType();
-            MilestoningStereotype targetTypeMilestoningStereotype = Milestoning.temporalStereotypes(target._stereotypes()).get(0);
+            MilestoningStereotype targetTypeMilestoningStereotype = Milestoning.temporalStereotypes(target._stereotypes());
             setMilestoningDates(property._parametersValues().toList(), targetTypeMilestoningStereotype, processingContext);
         }
         else
@@ -85,7 +103,7 @@ public class MilestoningDatePropagationHelper {
     public static void updateMilestoningPropagationContextForFilter(SimpleFunctionExpression func, ProcessingContext processingContext)
     {
         ValueSpecification parameterValue = func._parametersValues().getFirst();
-        if(parameterValue instanceof SimpleFunctionExpression && checkGetAllFunctionWithMilestoningContext((SimpleFunctionExpression) parameterValue))
+        if(parameterValue instanceof SimpleFunctionExpression && isGetAllFunctionWithMilestoningContext((SimpleFunctionExpression) parameterValue))
         {
             setMilestoningPropagationContext((SimpleFunctionExpression) parameterValue, processingContext);
         }
@@ -93,30 +111,30 @@ public class MilestoningDatePropagationHelper {
 
     public static void updateMilestoningPropagationContextForAutoMap(AbstractProperty<?> property, CompileContext context, int parametersCount, ValueSpecification processedParameter, ProcessingContext processingContext)
     {
-        if (Milestoning.isGeneratedMilestonedQualifiedPropertyWithMissingDates(property, context, parametersCount) && processedParameter instanceof SimpleFunctionExpression && ((SimpleFunctionExpression) processedParameter)._func() instanceof AbstractProperty && Milestoning.isGeneratedMilestoningProperty((AbstractProperty<?>) ((SimpleFunctionExpression) processedParameter)._func(), context))
+        if (isGeneratedMilestonedQualifiedPropertyWithMissingDates(property, context, parametersCount) && processedParameter instanceof SimpleFunctionExpression && ((SimpleFunctionExpression) processedParameter)._func() instanceof AbstractProperty && isGeneratedMilestoningProperty((AbstractProperty<?>) ((SimpleFunctionExpression) processedParameter)._func(), context))
         {
-            MilestoningDatePropagationHelper.updateMilestoningPropagationContext((SimpleFunctionExpression) processedParameter, processingContext);
+            updateMilestoningPropagationContext((SimpleFunctionExpression) processedParameter, processingContext);
         }
     }
 
     public static void updateMilestoningContext(AbstractProperty<?> property, ProcessingContext processingContext, CompileContext context, SimpleFunctionExpression func)
     {
-        if (Milestoning.isGeneratedQualifiedPropertyWithDatePropagationSupported(property, context))
+        if (isGeneratedQualifiedPropertyWithDatePropagationSupported(property, context))
         {
-            MilestoningDatePropagationHelper.updateMilestoningPropagationContext(func, processingContext);
+            updateMilestoningPropagationContext(func, processingContext);
         }
-        else if(property instanceof QualifiedProperty || property.getName().endsWith(Milestoning.ALL_VERSIONS_PROPERTY_NAME_SUFFIX) || property.getName().endsWith(Milestoning.ALL_VERSIONS_IN_RANGE_PROPERTY_NAME_SUFFIX))
+        else if(property instanceof QualifiedProperty || property.getName().endsWith(ALL_VERSIONS_PROPERTY_NAME_SUFFIX) || property.getName().endsWith(ALL_VERSIONS_IN_RANGE_PROPERTY_NAME_SUFFIX))
         {
             processingContext.milestoningDatePropagationContext.setProcessingDate(null);
             processingContext.milestoningDatePropagationContext.setBusinessDate(null);
         }
     }
 
-    public static void checkForValidSource(AppliedFunction appliedFunction, boolean isDatePropagationSupported, ProcessingContext processingContext)
+    public static void isValidSource(AppliedFunction appliedFunction, ProcessingContext processingContext)
     {
         if (!MILESTONING_DATE_SOURCE_TYPES.contains(appliedFunction.function) && !appliedFunction.parameters.isEmpty() && appliedFunction.parameters.get(0) instanceof AppliedFunction && ((AppliedFunction) appliedFunction.parameters.get(0)).function.equals("getAll"))
         {
-            processingContext.milestoningDatePropagationContext.isDatePropagationSupported = isDatePropagationSupported;
+            processingContext.milestoningDatePropagationContext.isDatePropagationSupported = false;
         }
     }
 
@@ -127,14 +145,205 @@ public class MilestoningDatePropagationHelper {
             processingContext.milestoningDatePropagationContext.setProcessingDate(null);
             processingContext.milestoningDatePropagationContext.setBusinessDate(null);
         }
-        if (result instanceof SimpleFunctionExpression && MilestoningDatePropagationHelper.checkGetAllFunctionWithMilestoningContext((SimpleFunctionExpression) result) && processingContext.milestoningDatePropagationContext.isDatePropagationSupported) {
-            MilestoningDatePropagationHelper.setMilestoningPropagationContext((SimpleFunctionExpression) result, processingContext);
+        if (result instanceof SimpleFunctionExpression && isGetAllFunctionWithMilestoningContext((SimpleFunctionExpression) result) && processingContext.milestoningDatePropagationContext.isDatePropagationSupported) {
+            setMilestoningPropagationContext((SimpleFunctionExpression) result, processingContext);
         }
-        if (result instanceof SimpleFunctionExpression && MilestoningDatePropagationHelper.checkForFilter((SimpleFunctionExpression) result))
+        if (result instanceof SimpleFunctionExpression && isFilter((SimpleFunctionExpression) result))
         {
-            MilestoningDatePropagationHelper.updateMilestoningPropagationContextForFilter((SimpleFunctionExpression) result, processingContext);
+            updateMilestoningPropagationContextForFilter((SimpleFunctionExpression) result, processingContext);
         }
     }
 
-    private static final ImmutableList<String> MILESTONING_DATE_SOURCE_TYPES = Lists.immutable.of("getAll", "exists", "project", "filter", "subType", "map");
+    public static boolean isAllVersionsInRangeProperty(AbstractProperty<?> property, CompileContext context)
+    {
+        return isGeneratedMilestoningProperty(property, context) && property._name().endsWith(ALL_VERSIONS_IN_RANGE_PROPERTY_NAME_SUFFIX);
+    }
+
+    public static boolean isGeneratedMilestoningProperty(AbstractProperty<?> property, CompileContext context)
+    {
+        String stereotype = String.valueOf(Milestoning.GeneratedMilestoningStereotype.generatedmilestoningproperty);
+        org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.extension.Profile profile = context.pureModel.getProfile("meta::pure::profiles::milestoning");
+        Stereotype milestoningStereotype = (Stereotype) Profile.findStereotype(profile, stereotype);
+        RichIterable<? extends Stereotype> stereotypes =  property._stereotypes();
+        return stereotypes.detect(s -> s != null && milestoningStereotype.equals(s)) != null;
+    }
+
+    public static boolean isGeneratedQualifiedPropertyWithDatePropagationSupported(AbstractProperty<?> property, CompileContext context)
+    {
+        return property instanceof QualifiedProperty && isGeneratedMilestoningProperty(property, context) && !isAllVersionsInRangeProperty(property, context);
+    }
+
+    private static int getCountOfParametersSatisfyingMilestoningDateRequirments(QualifiedProperty milestonedQualifiedProperty, CompileContext context)
+    {
+        if (!isGeneratedMilestoningProperty(milestonedQualifiedProperty, context))
+        {
+            throw new EngineException("Unable to get milestoning date parameters for non milestoned QualifiedProperty: " + milestonedQualifiedProperty.getName());
+        }
+        Class returnType = (Class)milestonedQualifiedProperty._genericType()._rawType();
+        MilestoningStereotype milestoningStereotype = Milestoning.temporalStereotypes(returnType._stereotypes());
+        return 1 + milestoningStereotype.getTemporalDatePropertyNames().size();
+    }
+
+    public static boolean isGeneratedMilestonedQualifiedPropertyWithMissingDates(AbstractProperty<?> property, CompileContext context, Integer parametersCount)
+    {
+        if (isGeneratedQualifiedPropertyWithDatePropagationSupported(property, context))
+        {
+            return parametersCount != getCountOfParametersSatisfyingMilestoningDateRequirments((QualifiedProperty)property, context);
+        }
+        return false;
+    }
+
+    public static boolean isProcessingTemporal(MilestoningStereotype milestoningStereotype)
+    {
+        return milestoningStereotype != null && milestoningStereotype.getPurePlatformStereotypeName() == "processingtemporal";
+    }
+
+    public static boolean isBusinessTemporal(MilestoningStereotype milestoningStereotype)
+    {
+        return milestoningStereotype != null && milestoningStereotype.getPurePlatformStereotypeName() == "businesstemporal";
+    }
+
+    private static boolean isSingleDateTemporal(MilestoningStereotype milestoningStereotype)
+    {
+        return milestoningStereotype != null && isProcessingTemporal(milestoningStereotype) || isBusinessTemporal(milestoningStereotype);
+    }
+
+    private static Type getMilestonedPropertyOwningType(AbstractProperty<?>property)
+    {
+        if (property._owner() instanceof Class)
+        {
+            return (Class) property._owner();
+        }
+        else if (property._owner() instanceof org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.relationship.Association)
+        {
+            return ((org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.relationship.Association) property._owner())._originalMilestonedProperties().toList().select(prop -> prop._name() != property._name()).getFirst()._genericType()._rawType();
+
+        }
+        return null;
+    }
+
+    private static Pair<MilestoningStereotype, MilestoningStereotype> getSourceTargetMilestoningStereotypes(AbstractProperty<?> property)
+    {
+        Class source = (Class) getMilestonedPropertyOwningType(property);
+        Class target = (Class) property._genericType()._rawType();
+        MilestoningStereotype sourceTypeMilestoningStereotype = Milestoning.temporalStereotypes(source._stereotypes()) != null ? Milestoning.temporalStereotypes(source._stereotypes()) : null;
+        MilestoningStereotype targetTypeMilestoningStereotype = Milestoning.temporalStereotypes(target._stereotypes()) != null ? Milestoning.temporalStereotypes(target._stereotypes()) : null;
+        return Tuples.pair(sourceTypeMilestoningStereotype, targetTypeMilestoningStereotype);
+    }
+
+    public static boolean isBiTemporal(MilestoningStereotype milestoningStereotype)
+    {
+        return milestoningStereotype != null && milestoningStereotype.getPurePlatformStereotypeName() == "bitemporal";
+    }
+
+    private static void setMilestoningDateParameters(ValueSpecification[] dateParamValues, int index, ValueSpecification milestoningDate)
+    {
+        dateParamValues[index] = milestoningDate;
+    }
+
+    private static boolean oneDateParamSupplied(ListIterable<? extends ValueSpecification> parameterValues)
+    {
+        return parameterValues.size() == 2;
+    }
+
+    private static boolean noDateParamSupplied(ListIterable<? extends ValueSpecification> parameterValues)
+    {
+        return parameterValues.size() == 1;
+    }
+
+    public static void applyPropertyFunctionExpressionMilestonedDates(FunctionExpression fe, AbstractProperty<?> func, SourceInformation sourceInformation, ProcessingContext processingContext)
+    {
+        Pair<MilestoningStereotype, MilestoningStereotype> sourceTargetMilestoningStereotypes = getSourceTargetMilestoningStereotypes(func);
+        MilestoningStereotype sourceTypeMilestoning = sourceTargetMilestoningStereotypes.getOne();
+        MilestoningStereotype targetTypeMilestoning = sourceTargetMilestoningStereotypes.getTwo();
+        String propertyName = func.getName();
+        MutableList<? extends ValueSpecification> parametersValues = fe._parametersValues().toList();
+        ValueSpecification[] milestoningDateParameters = new ValueSpecification[targetTypeMilestoning.getTemporalDatePropertyNames().size()];
+        fe._originalMilestonedPropertyParametersValues(fe._parametersValues());
+
+        if (isBiTemporal(targetTypeMilestoning))
+        {
+            if (isBiTemporal(sourceTypeMilestoning) && oneDateParamSupplied(parametersValues))
+            {
+                milestoningDateParameters[0] =  processingContext.milestoningDatePropagationContext.getProcessingDate();
+                milestoningDateParameters[1] = parametersValues.get(1);
+            }
+            else if (isSingleDateTemporal(sourceTypeMilestoning) && oneDateParamSupplied(parametersValues))
+            {
+                int propagatedDateIndex = Objects.requireNonNull(sourceTypeMilestoning).positionInTemporalParameterValues();
+                ValueSpecification propagatedDate;
+                int otherPropagatedDateIndex;
+                if (isProcessingTemporal(sourceTypeMilestoning))
+                {
+                    propagatedDate =  processingContext.milestoningDatePropagationContext.getProcessingDate();
+                    otherPropagatedDateIndex = 1;
+                }
+                else
+                {
+                    propagatedDate =  processingContext.milestoningDatePropagationContext.getBusinessDate();
+                    otherPropagatedDateIndex = 0;
+                }
+                setMilestoningDateParameters(milestoningDateParameters, propagatedDateIndex, propagatedDate);
+                setMilestoningDateParameters(milestoningDateParameters, otherPropagatedDateIndex, parametersValues.get(1));
+            }
+            if (isBiTemporal(sourceTypeMilestoning) && noDateParamSupplied(parametersValues))
+            {
+                milestoningDateParameters[0] = processingContext.milestoningDatePropagationContext.getProcessingDate();
+                milestoningDateParameters[1] = processingContext.milestoningDatePropagationContext.getBusinessDate();
+            }
+            if(milestoningDateParameters[0] == null || milestoningDateParameters[1] == null)
+            {
+                throw new EngineException("No-Arg milestoned property: '" + propertyName + "' must be either called in a milestoning context or supplied with " + "[processingDate, businessDate]" + " parameters", sourceInformation, EngineErrorType.COMPILATION);
+            }
+        }
+        else if (isSingleDateTemporal(targetTypeMilestoning) && noDateParamSupplied(parametersValues))
+        {
+            ValueSpecification propagatedDate;
+            if (isProcessingTemporal(targetTypeMilestoning))
+            {
+                propagatedDate = processingContext.milestoningDatePropagationContext.getProcessingDate();
+            }
+            else
+            {
+                propagatedDate = processingContext.milestoningDatePropagationContext.getBusinessDate();
+            }
+            if (isBiTemporal(sourceTypeMilestoning))
+            {
+                setMilestoningDateParameters(milestoningDateParameters, 0, propagatedDate);
+            }
+            if (sourceTypeMilestoning == targetTypeMilestoning)
+            {
+                setMilestoningDateParameters(milestoningDateParameters, 0, propagatedDate);
+            }
+            if (milestoningDateParameters[0] == null)
+            {
+                if (isProcessingTemporal(targetTypeMilestoning))
+                {
+                    throw new EngineException("No-Arg milestoned property: '" + propertyName + "' must be either called in a milestoning context or supplied with " + "[processingDate]" + " parameters", sourceInformation, EngineErrorType.COMPILATION);
+                }
+                else
+                {
+                    throw new EngineException("No-Arg milestoned property: '" + propertyName + "' must be either called in a milestoning context or supplied with " + "[businessDate]" + " parameters", sourceInformation, EngineErrorType.COMPILATION);
+                }
+            }
+        }
+        if (!ArrayIterate.isEmpty(milestoningDateParameters))
+        {
+            parametersValues = LazyIterate.concatenate(FastList.<ValueSpecification>newListWith(parametersValues.get(0)), FastList.newListWith(milestoningDateParameters)).toList();
+            fe._parametersValues(parametersValues);
+        }
+    }
+
+    public static void updateFunctionExpressionWithMilestoningDateParams(FunctionExpression functionExpression, AbstractProperty<?> propertyFunc, SourceInformation sourceInformation, ProcessingContext processingContext)
+    {
+        applyPropertyFunctionExpressionMilestonedDates(functionExpression, propertyFunc, sourceInformation, processingContext);
+        String propertyName = propertyFunc._name();
+        Class owner = (Class) getMilestonedPropertyOwningType(propertyFunc);
+        Object prop = ListIterate.select(owner._qualifiedProperties().toList(), p -> p instanceof QualifiedProperty && ((QualifiedProperty) p)._name() == propertyName).getFirst();
+        if (prop == null)
+        {
+            prop = ListIterate.select(owner._qualifiedPropertiesFromAssociations().toList(), p -> p instanceof QualifiedProperty && ((QualifiedProperty) p)._name() == propertyName).getFirst();
+        }
+        functionExpression._func((org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.function.Function<? extends Object>) prop);
+    }
 }
